requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic[email]
openai>=1.1.0
python-dotenv>=1.0.0
gradio>=4.0.0
pypdf>=3.17.0
requests>=2.31.0
email-validator>=2.1.0
numpy>=1.24.0
scikit-learn>=1.0.0